requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
bleach==6.0.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cron-descriptor==1.4.0
Django==4.2.1
django-admin-geomap==1.5
django-celery-beat==2.5.0
django-constance==2.9.1
django-filter==23.2
django-summernote==0.8.20.0
django-timezone-field==5.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
kombu==5.2.4
Markdown==3.4.3
openpyxl==3.1.2
Pillow==9.5.0
prompt-toolkit==3.0.38
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
XlsxWriter==3.1.0
